requirements.txt
selenium>=4.10.0
beautifulsoup4>=4.12.2
webdriver-manager>=3.8.6
httpx>=0.27.0
python-dotenv>=1.1.0